third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/event-listener-breakpoints-expected.txt

Tests event listener breakpoints.


Running: testClickBreakpoint
Script execution paused.
Call stack:
    0) testElementEventListener (event-listener-breakpoints.js:16)
    1) addListenerAndClick (event-listener-breakpoints.js:23)
    2)  (:1)
Event target: INPUT
Script execution resumed.

Running: testAuxclickBreakpoint
Script execution paused.
Call stack:
    0) testElementEventListener (event-listener-breakpoints.js:16)
    1) addListenerAndAuxclick (event-listener-breakpoints.js:30)
    2)  (:1)
Event target: INPUT
Script execution resumed.

Running: testTimerFiredBreakpoint
Script execution paused.
Call stack:
    0) timerFired (event-listener-breakpoints.js:42)
Script execution resumed.

Running: testLoadBreakpointOnXHR
Script execution paused.
Call stack:
    0) loadCallback (event-listener-breakpoints.js:65)
Event target: XMLHttpRequest
Script execution resumed.

Running: testMediaEventBreakpoint
Script execution paused.
Call stack:
    0) onVideoPlay (event-listener-breakpoints.js:77)
Event target: VIDEO
Script execution resumed.

Running: testPointerEventBreakpoint
Script execution paused.
Call stack:
    0) testElementEventListener (event-listener-breakpoints.js:16)
    1) addListenerAndPointerDown (event-listener-breakpoints.js:37)
    2)  (:1)
Event target: INPUT
Script execution resumed.
